requirements.txt
requests
dnspython
whois
Pillow
colorama
